MYSQL connection added
Added code to python file that connects the python file to the MYSQL connection. Also added code to print all the databases found in the MYSQL connection. 

On the JavaScript file, I tried to figure out ways to assign data passed from the python file (flask) to a JavaScript file. In other words, if the python file receives an address, finds ward 4 associated with that address in the database, and finally it passes the 4 back to the JavaScript and we assign that 4 to a JavaScript variable.