Improved security for Python file
Previously, our database credentials were hardcoded into the python file. The database password has been changed, and the credentials have been taken out of the code on 'test_server.py'. The credentials now reside in a file called secrets.py, that will never be pushed to GitHub. To use these credentials, just put 'from secrets import dataBase' in the test_server.py file, and you will now have the variable you need to interact with the database. This was done to improve security measures, due to the fact that hardcoding sensitive information such as database credentials or API keys is a bad practice.